% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_parallel.R
\name{download_objects}
\alias{download_objects}
\title{Download objects from an MN, optionally in parallel}
\usage{
download_objects(node, pids, path = getwd(), ext = ".xml", overwrite = FALSE)
}
\arguments{
\item{node}{(CNode|MNode) The Node to download from}

\item{pids}{(character) One ore more PIDs to download from the \code{node}}

\item{path}{(character) The destination directory for downloadd Objects}

\item{ext}{(character) Optional. Specify a custom file extension to add to each downloaded file}

\item{overwrite}{(boolean) Whether to overwrite existing files (TRUE) or not (FALSE)}
}
\value{
(list) A list of file paths for the downloaded files
}
\description{
This function uses the \code{future} package and therefore allows
downloads to happen in sequential order or parallel.
}
