% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ea.R
\name{ea_to_csv}
\alias{ea_to_csv}
\title{Extract Entity-Attribute information from a set of EML records}
\usage{
ea_to_csv(paths, datetime = Sys.time())
}
\arguments{
\item{paths}{(character) Paths to one or more EML records to read from}

\item{datetime}{(optional) Optional. Specify a query time to save with the
resulting CSV}
}
\value{
(data.frame) A data.frame of Entity-Attribute information
}
\description{
Extract Entity-Attribute information from a set of EML records
}
