% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ea.R
\name{extract_ea}
\alias{extract_ea}
\title{Extract Entity-Attribute information from a set of EML records}
\usage{
extract_ea(paths, datetime = Sys.time())
}
\arguments{
\item{paths}{(character) Paths to one or more EML records to read from}

\item{datetime}{(optional) Optional. Specify a query time to save with the
resulting CSV}
}
\value{
(data.frame) A data.frame of Entity-Attribute information
}
\description{
Extract Entity-Attribute information from a set of EML records
}
