% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_keywords.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract keywords from EML records}
\usage{
extract_keywords(paths, datetime = Sys.time())
}
\arguments{
\item{paths}{(character) Paths to one or more EML records to read from}

\item{datetime}{(optional) Optional. Specify a query time to save with the
resulting CSV}
}
\value{
(data.frame) A data.frame of keywords, one row per keyword
}
\description{
Extract keywords from EML records
}
