% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_ea.R
\name{get_object_ea}
\alias{get_object_ea}
\title{Get Entity-Attribute Metadata for one or more Objects}
\usage{
get_object_ea(node, identifiers)
}
\arguments{
\item{node}{(MNode|CNode) The Node where the Object(s) can be found}

\item{identifiers}{The Object's identifier (PID)}
}
\value{
(data.frame) A table of entity attribute metadata
}
\description{
Get Entity-Attribute Metadata for one or more Objects
}
