% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_parallel.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Santize filenames into valid filesystem paths}
\usage{
sanitize_filename(filenames)
}
\arguments{
\item{filenames}{(character) One or more filenames to sanitized}
}
\value{
(character) The sanitized filenames
}
\description{
This is a custom routine that's probably overly aggressive
}
